What I had when I started this repo
I was originally working on a hash table project on my computer. However, once I realized how much there was to this project, I decided to make it a repo. This is what I had when I made that decision.

catch.hpp is just for unit testing (see Test.cxx).