Changed Increment_Bucket_Index to Calculate_Bucket_Index
I totally redid the way that the new bucket index is calculated. the function now accepts the hash index number of buckets searched as arguments and returns the new bucket index.

this function is easier to work with (no returning via arguments) and is more general. It can easily be changed to use general linear or quadratic probing (though double hashing would be different).

I also updated a bunch of comments to reflect the changes.

The updated code appears to run just like the old code did (though it's easier to understand what's going on now + the code is more general!)