Added resizing + caught a bug
I added a resizing method to the code (+ added several new tests)

I had to modify the resizing method to work with chaining (the original one was written for open aliasing).

Doing this made me change two things: Firstly, I had to make the hash tablr class a freind of the item list one. Secondly, I found a bug in Item List's remove method.

The old Put method broke the list if the item being removed happened to be the last item in the list. This is because it failed to update the end pointer (which is used by the put method). Tbh, I don't know why this bug wasn't crashing my code before. Either way, I fixed it. See lines 150-159 of HashTable.cxx

The code runs and passes all tests!