Split off Item_List into its own file + Added compile guards
The main HashTable.cxx was getting very cluttered. I split the Item_List part of the code into its own file (along with the Item structure and Item_Node class).

I also added compile guards to both HashTable.cxx and ItemList.hxx

In the next commit, I plan on renaming HashTable.cxx to HashTable.hxx. I didn't do that in this commit just so that the changes to HashTable.cxx are visible.