Fixed a BIG issue in removal
The Removal function had a catastrophic bug in it. When an item was removed, we deleted the entry and then set entry to entry->getNext. The issue is that the "entry" in that second statement is deleted and therefore shouldn't work. Idk how I missed this  before or how it didn't crash my code. Either way, I fixed it. See Lines:170-180 of ItemList.hxx.... basically, if an item is removed, then Prev is unchanged and Entry goes to the next item. Otherwise, prev and entry are updated the usual way.